Make pyramid_zipkin faster when the request is not sampled
If the request is not sampled, there's no reason to wrap the handler
with py_zipkin's zipkin_span handler. Also, I've added a new config that
allows services to not add all zipkin headers if they're really
concerned with performance.

This should make pyramid_zipkin about 500us faster when the request is
not sasmpled.