Allow using the pyramid route pattern as span name.
Using the url path as span name was a bad idea. Urls often have ids in
the path, which make the span names very high cardinality.
This makes it impossible to filter by span name in the zipkin UI and
adds pressure on the cassandra indices. The full url is anyway added as
a tag, so that information is not lost.

This defaults to False to maintain backward compatibility.