Use direct assignment rather then set_property
request.set_property is relatively slow (~100 us) and it's an overkill
in our case. According to the docs it's useful since the callback is
evaluated when the property is accessed the first time rather than when
you call set_property. That saves time and cpu in case this property is
never used but:
1. we read it the next line, so we pay the penality anyway
2. this seems useful for very slow callback, while our returns almost
   immediately

Directly assigning zipkin_trace_id decreases the set time from ~100us to
~10us and the read time from ~10us to ~1us.